refactor: move tests to a root dir
